+++ source/enemies/bloodtailhawk.js
@@ -1,6 +1,7 @@
 const { EnemyTemplate } = require("../classes");
 const { selectRandomFoe, nextRepeat } = require("../shared/actionComponents");
 const { dealDamage } = require("../util/combatantUtil");
+const { getEmoji } = require("../util/elementUtil");
 
 module.exports = new EnemyTemplate("Bloodtail Hawk",
 	"Wind",
@@ -13,6 +14,7 @@ module.exports = new EnemyTemplate("Bloodtail Hawk",
 ).addAction({
 	name: "Rake",
 	element: "Wind",
+	description: `Deals ${getEmoji("Wind")} damage to a single target`,
 	priority: 0,
 	effect: ([target], user, isCrit, adventure) => {
 		let damage = 45;