+++ source/enemies/clone.js
@@ -8,4 +8,9 @@ module.exports = new EnemyTemplate("@{clone}",
 	0,
 	"clone", // this shouldn't get used, clones always copy delvers
 	true
-).setPower(35);
+).setPower(35)
+	.addAction({
+		name: "Mirroring Moves",
+		element: "Untyped",
+		description: "Clones will use the same move on a mirrored target as the delver they've copied."
+	});