+++ source/enemies/firearrowfrog.js
@@ -1,6 +1,7 @@
 const { EnemyTemplate } = require("../classes");
 const { selectRandomFoe, selectSelf } = require("../shared/actionComponents.js");
 const { addModifier, dealDamage } = require("../util/combatantUtil");
+const { getEmoji } = require("../util/elementUtil.js");
 
 const PATTERN = {
 	"Venom Cannon": "random",
@@ -22,6 +23,7 @@ module.exports = new EnemyTemplate("Fire-Arrow Frog",
 ).addAction({
 	name: "Venom Cannon",
 	element: "Fire",
+	description: `Inflict minor ${getEmoji("Fire")} damage and Poison on a single foe`,
 	priority: 0,
 	effect: ([target], user, isCrit, adventure) => {
 		let damage = 20;
@@ -38,6 +40,7 @@ module.exports = new EnemyTemplate("Fire-Arrow Frog",
 }).addAction({
 	name: "Burrow",
 	element: "Untyped",
+	description: "Gain Evade",
 	priority: 0,
 	effect: (targets, user, isCrit, adventure) => {
 		let stacks = 2;
@@ -54,6 +57,7 @@ module.exports = new EnemyTemplate("Fire-Arrow Frog",
 }).addAction({
 	name: "Goop Spray",
 	element: "Untyped",
+	description: "Slow a single foe",
 	priority: 0,
 	effect: ([target], user, isCrit, adventure) => {
 		if (isCrit) {