+++ ChangeLog.md
@@ -20,6 +20,7 @@
 - Buffed Elkemist ||Toil removes a random debuff, Bubble converts buffs to Fire Weakness, Progress no longer grants Stasis||
 - Forges are now Workshops: They now always offer repairs and upgrades. They will randomly offer one of the following: changing gear upgrades, increasing party gear capacity, or trading for an unknown upgraded gear.
 - Gear that applies Stagger now accounts for Same Element Stagger Bonus in their descriptions
+- New Command: `/manual enemy-info`
 ## Prophets of the Labyrinth v0.12.0:
 - Reintroduced Light and Darkness elements
 - Reworked Martial Artist starting gear